Efficient source selection for perm True

+++ test/test_batched_bss_float32.py
@@ -65,7 +65,7 @@ class TestBatchedBSS(unittest.TestCase):
         self.assertTrue(torch.allclose(sdr, sdr_b[0]))
         self.assertTrue(torch.allclose(sir, sir_b[0]))
         self.assertTrue(torch.allclose(sar, sar_b[0]))
-        self.assertTrue(torch.allclose(perm, perm_b[0]))
+        self.assertTrue(torch.allclose(perm.long(), perm_b[0]))
 
     @unittest.skipIf(BACKPROP is False, 'System non backpropagable')
     def test_bss_eval_gradient_flow(self):
