+++ test/test_batched_bss_float64.py
@@ -65,4 +65,4 @@ class TestBatchedBSS(unittest.TestCase):
         self.assertTrue(torch.allclose(sdr, sdr_b[0]))
         self.assertTrue(torch.allclose(sir, sir_b[0]))
         self.assertTrue(torch.allclose(sar, sar_b[0]))
-        self.assertTrue(torch.allclose(perm, perm_b[0]))
+        self.assertTrue(torch.allclose(perm.long(), perm_b[0]))
